backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "pablomigueldias",email = "pablo.miguel.dias@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.128.1,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "beanie (>=2.0.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (==4.0.1)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"